requirements.txt
torch==2.4.0
torchvision==0.19.0
typer==0.15.1
fastapi==0.115.6
uvicorn==0.33.0
requests==2.32.3
python-multipart==0.0.20
omegaconf==2.3.0
hydra-core==1.3
pandas==2.2.3
seaborn==0.13.2
opencv-python==4.10.0.84
imutils==0.5.4
kagglehub==0.3.6
torch_tb_profiler==0.4.3
loguru==0.7.3
pytorch-lightning==2.5.0.post0
wandb==0.19.2
dvc-gs==3.0.1
invoke==2.2.0
onnx==1.17.0
onnxruntime==1.19.2
onnxscript==0.1.0.dev20250120
streamlit==1.41.1
google-cloud-run>=0.10.5
